--- migration/src/main/java/com/github/gfx/android/orma/migration/MigrationAbortException.java
@@ -1,8 +1,0 @@
-package com.github.gfx.android.orma.migration;
-
-public class MigrationAbortException extends RuntimeException {
-
-    public MigrationAbortException(Throwable throwable) {
-        super(throwable);
-    }
-}
